the footer disclaimer text colour has been enhanced
