+++ package.json
@@ -102,7 +102,6 @@
     "@vitejs/plugin-react": "^4.3.4",
     "@vitest/ui": "^1.6.0",
     "autoprefixer": "^10.4.16",
-    "baseline-browser-mapping": "^2.10.0",
     "danger": "^13.0.5",
     "eslint": "^10.2.1",
     "eslint-plugin-react-hooks": "^7.1.1",
