+++ src/services/walletPass.ts
@@ -1,7 +1,7 @@
 import {
   buildGoldPassPayload as buildSharedGoldPassPayload,
   buildWalletCardContent,
-} from "../../api/shared/walletPassModel.js";
+} from "../shared/walletPassModel.js";
 
 const HUSHH_WALLET_ENDPOINT = "/api/wallet-pass";
 const HUSHH_GOOGLE_WALLET_ENDPOINT = "/api/google-wallet-pass";
